refactor(message-type): rename, use in enquiries
